Refac: tests
