+++ tests/unit/config/lightning/test_optimizer_configs.py
@@ -6,7 +6,6 @@ import pytest
 from careamics.config.lightning.optimizer_configs import (
     LrSchedulerConfig,
     OptimizerConfig,
-    get_unknown_parameters,
 )
 from careamics.config.support.supported_optimizers import (
     SupportedOptimizer,
@@ -17,32 +16,14 @@ from careamics.config.support.supported_optimizers import (
 OPT = [opt.value for opt in SupportedOptimizer]
 SGD = [SupportedOptimizer.SGD.value]
 
-#
+# StepLR requires `step_size`
 SCH_NO_STEP = [
     sch.value for sch in SupportedScheduler if sch != SupportedScheduler.STEP_LR
 ]
 STEP = [SupportedScheduler.STEP_LR.value]
 
-#
-
-
-def a_test_func(param_1: str, param_2: int, param_3: float = 0.1):
-    """Test function for _get_unknown_parameters."""
-    pass
-
 
 # ------------------------ Unit tests --------------------------
-
-
-def test_get_unknown_parameters():
-    """Test that _get_unknown_parameters returns the correct unknown parameters."""
-    user_params = {
-        "param_1": "test",
-        "param_3": 0.2,
-        "param_4": "unknown",
-    }
-    unknown_params = get_unknown_parameters(a_test_func, user_params)
-    assert unknown_params == {"param_4": "unknown"}
 
 
 @pytest.mark.parametrize(
