+++ tests/unit/config/lightning/test_training_configuration.py
